fixed some signed/unsigned warnings

+++ tuxbox/neutrino/src/driver/slotbuffer.c
@@ -1,5 +1,5 @@
 /*
- * $Header: /cvs/tuxbox/apps/tuxbox/neutrino/src/driver/slotbuffer.c,v 1.2 2006/03/07 15:59:44 zwen Exp $
+ * $Header: /cvs/tuxbox/apps/tuxbox/neutrino/src/driver/slotbuffer.c,v 1.3 2009/08/12 19:33:00 houdini Exp $
  *
  * (C) 2004 by thegoodguy <thegoodguy@berlios.de>
  *
@@ -51,7 +51,7 @@ size_t slotbuffer_remaining_read_space(slotbuffer_t * const slotbuffer)
 	delta = slotbuffer->write_pointer;
 	delta -= slotbuffer->read_pointer;
 
-	return (delta >= 0) ? delta : (delta + slotbuffer->size);
+	return (delta >= 0) ? delta : (ssize_t)(delta + slotbuffer->size);
 }
 
 size_t slotbuffer_remaining_write_space(slotbuffer_t * const slotbuffer)
@@ -61,7 +61,7 @@ size_t slotbuffer_remaining_write_space(slotbuffer_t * const slotbuffer)
 	delta = slotbuffer->read_pointer;
 	delta -= slotbuffer->write_pointer;
 
-	return ((delta > 0) ? delta : (delta + slotbuffer->size)) - 1;
+	return ((delta > 0) ? delta : (ssize_t)((delta + slotbuffer->size)) - 1);
 }
 
 size_t slotbuffer_remaining_continuous_read_space(slotbuffer_t * const slotbuffer)
@@ -73,7 +73,7 @@ size_t slotbuffer_remaining_continuous_read_space(slotbuffer_t * const slotbuffe
 	read_pointer  = slotbuffer->read_pointer;
 	delta        -= read_pointer;
 
-	return (delta >= 0) ? delta : (slotbuffer->size - read_pointer);
+	return (delta >= 0) ? delta : (ssize_t)(slotbuffer->size - read_pointer);
 }
 
 size_t slotbuffer_remaining_continuous_write_space(slotbuffer_t * const slotbuffer)
@@ -89,7 +89,7 @@ size_t slotbuffer_remaining_continuous_write_space(slotbuffer_t * const slotbuff
 	write_pointer  = slotbuffer->write_pointer;
 	delta         -= write_pointer;
 
-	return (delta > 0) ? (delta - 1) : (slotbuffer->size - write_pointer);
+	return (delta > 0) ? (delta - 1) : (ssize_t)(slotbuffer->size - write_pointer);
 }
 
 void slotbuffer_advance_read_pointer(slotbuffer_t * const slotbuffer, const size_t delta)
